Modified structure, fixed Pop Games and added barebones NewlyAddedGames :
Added some docs for palettes.
Changed places for a few things.
Added barebone struct for NewlyAddedGames.
