+++ src/pages/Gamehub-01/Gamehub.jsx
@@ -1,6 +1,6 @@
 import { createEffect } from "solid-js";
 import './Gamehub.css'
-import PopularGames from "./Gamehub-Components/Popular-Games-01/Popular-Games";
+import PopularGames from "./Gamehub-Components-01/Popular-Games-01/Popular-Games";
 
 function Gamehub() {
