+++ src/palette_color.md
@@ -1,0 +1,16 @@
+
+## Default Dark Purple.
+```css
+:root[data-theme="default-dark-purple"] {
+    --accent-color: rgba(187, 114, 160, 1); /* #bb72a0 */
+    --secondary-color: rgba(110, 53, 103, 1); /* #6e3567 */
+    --secondary-30-selected-color: rgba(110, 53, 103, 0.3); /* #6e3567 */
+    --non-selected-text-color: rgba(236, 224, 240, 0.5); /* #ece0f0 */
+    --primary-color: rgba(194, 158, 209, 1); /* #c29ed1 */
+    --secondary-20-color: rgba(110, 53, 103, 0.2); /* #6e3567 */
+    --text-color: rgba(236, 224, 240, 1); /* #ece0f0 */
+    --background-color: rgba(11, 7, 14, 1); /* #0b070e */
+}
+```
+
+## More to be added soon...
